migrate kofi endpoint
